wording in TypeExercises
Parametricity is misspelled twice.

I am suggesting section 5 to be renamed Algebra from Logic. While the slides point to equivalence of algebra and logic, to me the focus of the exercises in Section 5 are actually of algebraic nature. Is there any consideration to introduce another section that is more compelling about logic?